Checkbox and number
