+++ gradio/_simple_templates/simpledropdown.py
@@ -33,7 +33,9 @@ class SimpleDropdown(FormComponent):
         visible: bool = True,
         elem_id: str | None = None,
         elem_classes: list[str] | str | None = None,
-        **kwargs,
+        render: bool = True,
+        root_url: str | None = None,
+        _skip_init_processing: bool = False,
     ):
         """
         Parameters:
@@ -49,6 +51,9 @@ class SimpleDropdown(FormComponent):
             visible: If False, component will be hidden.
             elem_id: An optional string that is assigned as the id of this component in the HTML DOM. Can be used for targeting CSS styles.
             elem_classes: An optional list of strings that are assigned as the classes of this component in the HTML DOM. Can be used for targeting CSS styles.
+                    render: bool = True,
+            root_url: str | None = None,
+            _skip_init_processing: bool = False,
         """
         self.choices = (
             # Although we expect choices to be a list of lists, it can be a list of tuples if the Gradio app
@@ -69,7 +74,9 @@ class SimpleDropdown(FormComponent):
             elem_id=elem_id,
             elem_classes=elem_classes,
             value=value,
-            **kwargs,
+            render=render,
+            root_url=root_url,
+            _skip_init_processing=_skip_init_processing,
         )
 
     def api_info(self) -> dict[str, Any]:
