+++ js/number/Index.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
 	import type { Gradio } from "@gradio/utils";
-	import Number from "./shared/Number.svelte";
-	import { Block } from "@gradio/atoms";
+	import { Block, BlockTitle } from "@gradio/atoms";
 	import { StatusTracker } from "@gradio/statustracker";
 	import type { LoadingStatus } from "@gradio/statustracker";
+	import { afterUpdate, tick } from "svelte";
 
 	export let gradio: Gradio<{
 		change: never;
@@ -28,6 +28,29 @@
 	export let value_is_output = false;
 	export let step: number | null = null;
 	export let mode: "static" | "interactive";
+
+	function handle_change(): void {
+		if (!isNaN(value) && value !== null) {
+			gradio.dispatch("change");
+			if (!value_is_output) {
+				gradio.dispatch("input");
+			}
+		}
+	}
+	afterUpdate(() => {
+		value_is_output = false;
+	});
+
+	async function handle_keypress(e: KeyboardEvent): Promise<void> {
+		await tick();
+		if (e.key === "Enter") {
+			e.preventDefault();
+			gradio.dispatch("submit");
+		}
+	}
+
+	$: value, handle_change();
+	$: disabled = mode === "static";
 </script>
 
 <Block
@@ -44,21 +67,62 @@
 		i18n={gradio.i18n}
 		{...loading_status}
 	/>
-	<Number
-		bind:value
-		bind:value_is_output
-		{label}
-		{info}
-		{show_label}
-		{minimum}
-		{maximum}
-		{step}
-		{container}
-		disabled={mode === "static"}
-		on:change={() => gradio.dispatch("change")}
-		on:input={() => gradio.dispatch("input")}
-		on:submit={() => gradio.dispatch("submit")}
-		on:blur={() => gradio.dispatch("blur")}
-		on:focus={() => gradio.dispatch("focus")}
-	/>
+	<label class="block" class:container>
+		<BlockTitle {show_label} {info}>{label}</BlockTitle>
+		<input
+			aria-label={label}
+			type="number"
+			bind:value
+			min={minimum}
+			max={maximum}
+			{step}
+			on:keypress={handle_keypress}
+			on:blur={() => gradio.dispatch("blur")}
+			on:focus={() => gradio.dispatch("focus")}
+			{disabled}
+		/>
+	</label>
 </Block>
+
+
+<style>
+	label:not(.container),
+	label:not(.container) > input {
+		height: 100%;
+		border: none;
+	}
+	.container > input {
+		border: var(--input-border-width) solid var(--input-border-color);
+		border-radius: var(--input-radius);
+	}
+	input[type="number"] {
+		display: block;
+		position: relative;
+		outline: none !important;
+		box-shadow: var(--input-shadow);
+		background: var(--input-background-fill);
+		padding: var(--input-padding);
+		width: 100%;
+		color: var(--body-text-color);
+		font-size: var(--input-text-size);
+		line-height: var(--line-sm);
+	}
+	input:disabled {
+		-webkit-text-fill-color: var(--body-text-color);
+		-webkit-opacity: 1;
+		opacity: 1;
+	}
+
+	input:focus {
+		box-shadow: var(--input-shadow-focus);
+		border-color: var(--input-border-color-focus);
+	}
+
+	input::placeholder {
+		color: var(--input-placeholder-color);
+	}
+
+	input:out-of-range {
+		border: var(--input-border-width) solid var(--error-border-color);
+	}
+</style>
